Fixes in services, route engine and route test
Service: add flag to avoid merging entries
When we retrieve data from multiple commands, we combine them into
a single flattened list, and in the process, combine the fields
for entries with the same key. However, in Palo Alto firewalls, the
route output in case of ECMP repeats the prefix for each next hop.
This causes the merge logic to fail by reducing all ECMP paths into
just the last one.

To fix this problem, we added a new key, merge, which is set to False
in case of PA firewall routes. Its absence is assumed to mean merging
is fine, which is the old behavior before the presence of this key.

This fixes the ECMP issue with PA firewall routes.

svcparser: handle case where dict val is None
Routes: add OIFs to routes that do not have OIFs (e.g. panos)
Routes test: fix tests and regenerate output due to differently sorted after the route engine fix

Co-authored-by: Dinesh Dutt <dd.ps4u@gmail.com>
Signed-off-by: Claudio Usai <c.usai186@gmail.com>